requirements.txt
cyclonedx-python-lib
requests
jinja2
pandas
spdx-tools
weasyprint